+++ src/main/java/com/sparta/meeting_platform/MeetingPlatformApplication.java
@@ -2,7 +2,9 @@ package com.sparta.meeting_platform;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
 
 @EnableJpaAuditing
 @SpringBootApplication
